requirements.txt
google-genai>=1.0
networkx>=3.2
pyvis>=0.3.2
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
pydantic>=2.0
gradio>=4.0
numpy>=1.24
tqdm>=4.65
python-dotenv>=1.0
